requirements.txt
-e .
-r ./update_managed_prefix_lists/lambda/requirements.txt
aws-cdk.core
aws-cdk.aws_lambda
aws_cdk.aws_lambda_event_sources 